requirements.txt
flask
flask-cors
python-dotenv
sqlalchemy
psycopg2-binary
passlib[bcrypt]
PyJWT
eventlet
requests
